fix(nav): managed click on the icon but out of the menu
